+++ packages/env/src/web.ts
@@ -4,8 +4,8 @@ import { z } from "zod";
 export const env = createEnv({
   client: {
     VITE_ENABLE_MERCH: z.enum(["true", "false"]).default("false"),
+    VITE_GOOGLE_MAPS_API_KEY: z.string().optional(),
     VITE_MEDIA_URL: z.url().optional(),
-    VITE_RADAR_PUBLISHABLE_KEY: z.string().optional(),
     VITE_SENTRY_DSN: z.url().optional(),
     VITE_SENTRY_ENVIRONMENT: z.string().default("development"),
     VITE_SERVER_URL: z.url(),